+++ beetmoverscript/script.py
@@ -20,6 +20,7 @@ from scriptworker.utils import retry_async, raise_future_exceptions
 
 from beetmoverscript.constants import (
     MIME_MAP, RELEASE_BRANCHES, CACHE_CONTROL_MAXAGE, RELEASE_EXCLUDE,
+    TEMPLATE_KEY_PLATFORMS, NORMALIZED_BALROG_PLATFORMS
 )
 from beetmoverscript.task import (
     validate_task_schema, add_balrog_manifest_to_artifacts,
@@ -31,7 +32,8 @@ from beetmoverscript.utils import (
     load_json, get_hash, get_release_props, generate_beetmover_manifest,
     get_size, alter_unpretty_contents, matches_exclude,
     get_candidates_prefix, get_releases_prefix, get_creds, get_bucket_name,
-    is_release_action, is_promotion_action, get_partials_props
+    is_release_action, is_promotion_action, get_partials_props,
+    get_product_name
 )
 
 log = logging.getLogger(__name__)
@@ -51,7 +53,7 @@ async def push_to_nightly(context):
 
     # find release properties and make a copy in the artifacts directory
     release_props_file = get_initial_release_props_file(context)
-    context.release_props = get_release_props(release_props_file)
+    context.release_props = get_release_props(context, release_props_file)
 
     # generate beetmover mapping manifest
     mapping_manifest = generate_beetmover_manifest(context)
@@ -294,11 +296,11 @@ def generate_balrog_info(context, artifact_pretty_name, locale, destinations, fr
         data["from_buildid"] = from_buildid
         if is_promotion_action(context.action):
             partials = get_partials_props(context.task)
-            # TODO: improve search by a dedicated dict with buildid as key
             for p in partials.values():
                 if p['buildid'] == str(from_buildid):
                     data['previousVersion'] = p['previousVersion']
                     data['previousBuildNumber'] = p['previousBuildNumber']
+                    break
 
     return data
 
@@ -313,14 +315,17 @@ def enrich_balrog_manifest(context, locale):
                                  'http://download.cdn.mozilla.net/pub'])
 
     enrich_dict = {
-        "appName": release_props["appName"],
+        "appName": get_product_name(release_props['appName'],
+                                    TEMPLATE_KEY_PLATFORMS[release_props['stage_platform']]),
         "appVersion": release_props["appVersion"],
         "branch": release_props["branch"],
         "buildid": release_props["buildid"],
         "extVersion": release_props["appVersion"],
         "hashType": release_props["hashType"],
         "locale": locale if not locale == 'multi' else 'en-US',
-        "platform": release_props["stage_platform"],
+        # XXX: temporary fix for devedition until we solve 1424482
+        "platform": NORMALIZED_BALROG_PLATFORMS.get(release_props["stage_platform"],
+                                                    release_props["stage_platform"]),
         "url_replacements": url_replacements
     }
 
@@ -366,9 +371,10 @@ async def put(context, url, headers, abs_filename, session=None):
 
 # upload_to_s3 {{{1
 async def upload_to_s3(context, s3_key, path):
-    app = context.release_props['appName'].lower()
+    product = get_product_name(context.release_props['appName'].lower(),
+                               TEMPLATE_KEY_PLATFORMS[context.release_props['stage_platform']])
     api_kwargs = {
-        'Bucket': get_bucket_name(context, app),
+        'Bucket': get_bucket_name(context, product),
         'Key': s3_key,
         'ContentType': mimetypes.guess_type(path)[0]
     }