+++ beetmoverscript/test/test_utils.py
@@ -3,15 +3,14 @@ import pytest
 import tempfile
 
 from beetmoverscript.test import (context, get_fake_valid_task,
-                                  get_fake_balrog_props, get_fake_checksums_manifest,
-                                  get_fake_balrog_props_path)
+                                  get_fake_balrog_props, get_fake_checksums_manifest)
 import beetmoverscript.utils as butils
 from beetmoverscript.utils import (generate_beetmover_manifest, get_hash,
                                    write_json, generate_beetmover_template_args,
                                    write_file, is_release_action, is_promotion_action,
                                    get_release_props, get_partials_props,
                                    matches_exclude, get_candidates_prefix,
-                                   get_releases_prefix)
+                                   get_releases_prefix, get_product_name)
 from beetmoverscript.constants import HASH_BLOCK_SIZE
 
 assert context  # silence pyflakes
@@ -154,18 +153,58 @@ def test_is_action_release_or_promotion(action, release, promotion):
 
 
 # get_release_props {{{1
-def test_get_release_props():
-    expected_release_props = {
-        'appName': 'Fake',
-        'appVersion': '99.0a1',
-        'branch': 'mozilla-central',
-        'stage_platform': 'android-api-15',
-        'buildid': '20990205110000',
-        'hashType': 'sha512',
-        'platform': 'android-api-15'
+@pytest.mark.parametrize("taskjson,locale, relprops, expected", ((
+    'task.json', False, {
+        "platform": "android",
+        "stage_platform": "android-api-16"
+    }, {
+        "platform": "android-api-16",
+        "stage_platform": "android-api-16"
     }
-    release_props = get_release_props(get_fake_balrog_props_path())
-    assert release_props == expected_release_props
+), (
+    'task.json', True, {
+        "platform": "macosx64",
+    }, {
+        "platform": "mac",
+        "stage_platform": "macosx64"
+    }
+), (
+    'task.json', False, {
+        "platform": "linux64",
+        "stage_platform": "linux64"
+    }, {
+        "platform": "linux-x86_64",
+        "stage_platform": "linux64"
+    }
+), (
+    'task_devedition.json', False, {
+        "platform": "macosx64",
+        "stage_platform": "macosx64-devedition"
+    }, {
+        "platform": "mac",
+        "stage_platform": "macosx64-devedition"
+    }
+), (
+    'task_devedition.json', True, {
+        "platform": "win64",
+    }, {
+        "platform": "win64",
+        "stage_platform": "win64-devedition"
+    }
+)))
+def test_get_release_props(context, mocker, taskjson, locale, relprops, expected):
+    context.task = get_fake_valid_task(taskjson)
+    if locale:
+        context.task['payload']['locale'] = 'lang'
+
+    def fake_json(*args, **kwargs):
+        return {
+            'properties': relprops
+        }
+
+    dummy_file = None
+    mocker.patch.object(butils, 'load_json', new=fake_json)
+    assert get_release_props(context, dummy_file) == expected
 
 
 # get_partials_props {{{1
@@ -245,3 +284,21 @@ def test_matches_exclude(keyname, expected):
         r"^.*/metoo/.*$",
     ]
     assert matches_exclude(keyname, excludes) == expected
+
+
+# product_name {{{1
+@pytest.mark.parametrize("appName,tmpl_key,expected", ((
+    "firefox", "dummy", "firefox",
+), (
+    "firefox", "devedition", "devedition",
+), (
+    "Firefox", "devedition", "Devedition",
+), (
+    "Fennec", "dummy", "Fennec",
+), (
+    "Firefox", "dummy", "Firefox",
+), (
+    "fennec", "dummy", "fennec",
+)))
+def test_get_product_name(appName, tmpl_key, expected):
+    assert get_product_name(appName, tmpl_key) == expected